Changed the name of function for getting genotypes cox.out
created a similar function for cox.out which is used with snp.spike

Moved generate cluster to similar location in code across all xSurvCox functions

moved code for setting up output columns in plinkCoxSurve to same location as other functions.

Purpose of moving blocks around is to as much as possible align the code across functions so that we can create a single function that runs for all these types. Moving it now also allows us to perform isolated tests to make sure it doesn't break the code.