Turned all the xCoxSurv functions into wrappers that run the program using objects for each input type.
Created classes for each input type that holds all the parameters that were previously being sent to each xCoxSurv function, as well as providing a unique function for the load, process and write core functionality.

Replaced the xCoxSurv functions with a single function that receives objects of different classes. This enables objects for each input type to share code where applicable.

Deleted commented out code in runOnChunks.R

Fixed bug in getPlinkGenoData; gdsfile was incorrectly specified as an argument.